make test independent on util go file
