+++ token/jwk/fetch_keys_blackbox_test.go
@@ -8,12 +8,13 @@ import (
 	"github.com/fabric8-services/fabric8-common/httpsupport"
 	"github.com/fabric8-services/fabric8-common/resource"
 	testtoken "github.com/fabric8-services/fabric8-common/test/auth"
-	testconfiguration "github.com/fabric8-services/fabric8-common/test/configuration"
+	tokensupport "github.com/fabric8-services/fabric8-common/test/generated/token"
 	"github.com/fabric8-services/fabric8-common/test/recorder"
 	"github.com/fabric8-services/fabric8-common/token/jwk"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/fabric8-services/fabric8-common/configuration"
 )
 
 func TestFetchKeys(t *testing.T) {
@@ -25,7 +26,7 @@ func TestFetchKeys(t *testing.T) {
 		err := r.Stop()
 		require.NoError(t, err)
 	}()
-	config := testconfiguration.NewDefaultMockTokenManagerConfiguration(t)
+	config := NewDefaultMockTokenManagerConfiguration(t)
 	config.GetAuthServiceURLFunc = func() string {
 		return "https://auth-ok"
 	}
@@ -65,4 +66,18 @@ func TestFetchKeys(t *testing.T) {
 		})
 	})
 
+}
+
+// NewDefaultMockTokenManagerConfiguration initializes a new mock configuration for a token manager
+// functions can be overridden afterwards if needed
+func NewDefaultMockTokenManagerConfiguration(t *testing.T) *tokensupport.ManagerConfigurationMock {
+	config := tokensupport.NewManagerConfigurationMock(t)
+	config.GetAuthServiceURLFunc = func() string {
+		return "https://auth.prod-preview.openshift.io"
+	}
+
+	config.GetDevModePrivateKeyFunc = func() []byte {
+		return []byte(configuration.DevModeRsaPrivateKey)
+	}
+	return config
 }
