+++ token/token_blackbox_test.go
@@ -9,7 +9,6 @@ import (
 	"testing"
 
 	"github.com/fabric8-services/fabric8-common/test/auth"
-	testconfiguration "github.com/fabric8-services/fabric8-common/test/configuration"
 	tokensupport "github.com/fabric8-services/fabric8-common/test/generated/token"
 	testsuite "github.com/fabric8-services/fabric8-common/test/suite"
 	"github.com/fabric8-services/fabric8-common/token"
@@ -22,6 +21,7 @@ import (
 	"github.com/satori/go.uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/fabric8-services/fabric8-common/configuration"
 )
 
 type TokenManagerTestSuite struct {
@@ -37,7 +37,7 @@ func TestRunTokenManagerTestSuite(t *testing.T) {
 func (s *TokenManagerTestSuite) SetupSuite() {
 	s.UnitTestSuite.SetupSuite()
 
-	s.config = testconfiguration.NewDefaultMockTokenManagerConfiguration(s.T())
+	s.config = NewDefaultMockTokenManagerConfiguration(s.T())
 	var err error
 	s.tm, err = token.NewManager(s.config)
 	require.NoError(s.T(), err)
@@ -182,7 +182,7 @@ func (s *TokenManagerTestSuite) TestParseValidTokenOK() {
 }
 
 func (s *TokenManagerTestSuite) TestAuthServiceURL() {
-	config := testconfiguration.NewDefaultMockTokenManagerConfiguration(s.T())
+	config := NewDefaultMockTokenManagerConfiguration(s.T())
 
 	s.T().Run("OK if auth URL does not have trailing slash", func(t *testing.T) {
 		config.GetAuthServiceURLFunc = func() string {
@@ -419,4 +419,18 @@ func (s *TokenManagerTestSuite) TestServiceAccountToken() {
 		assert.Equal(t, "failed to obtain token from auth server \"http://authservice.unauthorized\": invalid Service Account ID or secret", err.Error())
 		assert.Empty(t, saToken)
 	})
+}
+
+// NewDefaultMockTokenManagerConfiguration initializes a new mock configuration for a token manager
+// functions can be overridden afterwards if needed
+func NewDefaultMockTokenManagerConfiguration(t *testing.T) *tokensupport.ManagerConfigurationMock {
+	config := tokensupport.NewManagerConfigurationMock(t)
+	config.GetAuthServiceURLFunc = func() string {
+		return "https://auth.prod-preview.openshift.io"
+	}
+
+	config.GetDevModePrivateKeyFunc = func() []byte {
+		return []byte(configuration.DevModeRsaPrivateKey)
+	}
+	return config
 }
