+++ token/token_whitebox_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/fabric8-services/fabric8-common/configuration"
-	testcfg "github.com/fabric8-services/fabric8-common/test/configuration"
 	testsuite "github.com/fabric8-services/fabric8-common/test/suite"
+	tokensupport "github.com/fabric8-services/fabric8-common/test/generated/token"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -22,7 +22,7 @@ type TestWhiteboxTokenSuite struct {
 }
 
 func (s *TestWhiteboxTokenSuite) TestDefaultManager() {
-	config := testcfg.NewDefaultMockTokenManagerConfiguration(s.T())
+	config := NewDefaultMockTokenManagerConfiguration(s.T())
 	config.GetAuthServiceURLFunc = func() string {
 		return "https://auth.prod-preview.openshift.io"
 	}
@@ -77,7 +77,7 @@ func assertDefaultManager(t *testing.T, config ManagerConfiguration) {
 
 func (s *TestWhiteboxTokenSuite) TestKeyLoaded() {
 	// given
-	config := testcfg.NewDefaultMockTokenManagerConfiguration(s.T())
+	config := NewDefaultMockTokenManagerConfiguration(s.T())
 	config.GetAuthServiceURLFunc = func() string {
 		return "https://auth.prod-preview.openshift.io"
 	}
@@ -107,4 +107,18 @@ func (s *TestWhiteboxTokenSuite) TestKeyLoaded() {
 		// then
 		assert.Nil(t, key)
 	})
+}
+
+// NewDefaultMockTokenManagerConfiguration initializes a new mock configuration for a token manager
+// functions can be overridden afterwards if needed
+func NewDefaultMockTokenManagerConfiguration(t *testing.T) *tokensupport.ManagerConfigurationMock {
+	config := tokensupport.NewManagerConfigurationMock(t)
+	config.GetAuthServiceURLFunc = func() string {
+		return "https://auth.prod-preview.openshift.io"
+	}
+
+	config.GetDevModePrivateKeyFunc = func() []byte {
+		return []byte(configuration.DevModeRsaPrivateKey)
+	}
+	return config
 }
